\name{impel-package}
\alias{impel-package}
\alias{impel}
\docType{package}
\title{
\packageTitle{impel}
}
\description{
A model constructed by DNA methylation of enhancer region, a model constructed by RNA expression of target genes, and a model integrating above two models. These three models could be used to predict the recurrence of early lung cancer.
}
\author{
Yulan Deng
Maintainer: Yulan Deng
}
\references{
Identification of enhancer region DNA methylation as a biomarker for the prediction of postoperative recurrence in early non-small-cell lung cancer. Senyi deng, Yulan Deng, Nan Chen, Chenglin Guo, Yutao Su, Xinyin Lv, Chengwu Liu, Lin Ma, Feng Lin, Yaohui Chen, Qiang Pu Jiandong Mei, Lunxu Liu. 
}
\keyword{ package }
