\name{methylation_matrix}
\alias{methylation_matrix}
\docType{data}
\title{
A toy example of DNA methylation matrix
}
\description{
A DNA methylation matrix contains locus (in rows) and sample (in columns). 
The annotation of methylation locus was based on the hg38 reference genome.
}
\usage{data("methylation_matrix")}
\keyword{datasets}
